🧩 DQM6AW rework: split CI hotspots
