+++ .agentplane/tasks/202607260007-DQM6AW/pr/github-title.txt
@@ -1,1 +1,1 @@
-✅ DQM6AW task: Prepare semantic conflict rework routes [202607260007-DQM6AW]
+🚧 DQM6AW task: Prepare semantic conflict rework routes [202607260007-DQM6AW]
