+++ packages/agentplane/src/commands/pr/integrate/cmd.test.ts
@@ -821,225 +821,4 @@ describe("pr/integrate/cmd", () => {
       else process.env.GITHUB_TOKEN = originalGithubToken;
     }
   });
-
-  it("refuses protected-base integration when live PR identity cannot be confirmed", async () => {
-    mocks.requireOpenGithubPrAtHead.mockRejectedValue(
-      new CliError({
-        exitCode: 3,
-        code: "E_VALIDATION",
-        message: "GitHub PR was not found for task/T-1",
-      }),
-    );
-    mocks.prepareIntegrate.mockResolvedValue({
-      ctx: {
-        config: { paths: { workflow_dir: ".agentplane/tasks" } },
-        git: {},
-        taskBackend: {},
-        resolvedProject: { gitRoot: "/repo" },
-      },
-      resolved: { gitRoot: "/repo" },
-      loadedConfig: {
-        workflow_mode: "branch_pr",
-        paths: {
-          worktrees_dir: ".agentplane/worktrees",
-          workflow_dir: ".agentplane/tasks",
-        },
-        commit: { generic_tokens: [] },
-      },
-      task: { id: "T-1", title: "Task", tags: [], verify: [], status: "DOING" },
-      prDir: "/repo/.agentplane/tasks/T-1/pr",
-      metaPath: "/repo/.agentplane/tasks/T-1/pr/meta.json",
-      diffstatPath: "/repo/.agentplane/tasks/T-1/pr/diffstat.txt",
-      verifyLogPath: "/repo/.agentplane/tasks/T-1/pr/verify.log",
-      metaSource: {
-        base: "main",
-        branch: "task/T-1",
-        head_sha: "head-sha",
-      },
-      branch: "task/T-1",
-      base: "main",
-      verifyLogText: "",
-      branchHeadSha: "head-sha",
-      baseHeadSha: "base-sha",
-      changedPaths: [],
-      verifyCommands: [],
-      alreadyVerifiedSha: null,
-      shouldRunVerify: false,
-      protectedBaseRequiresPrMerge: true,
-    });
-    const { cmdIntegrate } = await import("./cmd.js");
-
-    const caught = await cmdIntegrate({
-      cwd: "/repo",
-      taskId: "T-1",
-      mergeStrategy: "squash",
-      runVerify: false,
-      dryRun: false,
-      quiet: false,
-    }).catch((err: unknown) => err);
-
-    expect(caught).toMatchObject({ code: "E_VALIDATION" });
-    expect(mocks.buildTaskHandoffArtifact).not.toHaveBeenCalled();
-    expect(mocks.writeTaskHandoff).not.toHaveBeenCalled();
-    expect(mocks.runSquashMerge).not.toHaveBeenCalled();
-    expect(mocks.finalizeIntegrate).not.toHaveBeenCalled();
-  });
-
-  it("refuses protected-base handoff when GitHub omits the observed base SHA", async () => {
-    mocks.requireOpenGithubPrAtHead.mockResolvedValue({
-      prNumber: 338,
-      prUrl: "https://github.com/example/repo/pull/338",
-      status: "OPEN",
-      mergedAt: null,
-      mergeCommit: null,
-      base: "main",
-      baseSha: null,
-      headSha: "head-sha",
-    });
-    mocks.prepareIntegrate.mockResolvedValue({
-      ctx: {
-        config: { paths: { workflow_dir: ".agentplane/tasks" } },
-        git: {},
-        taskBackend: {},
-        resolvedProject: { gitRoot: "/repo" },
-      },
-      resolved: { gitRoot: "/repo" },
-      loadedConfig: {
-        workflow_mode: "branch_pr",
-        paths: {
-          worktrees_dir: ".agentplane/worktrees",
-          workflow_dir: ".agentplane/tasks",
-        },
-        commit: { generic_tokens: [] },
-      },
-      task: { id: "T-1", title: "Task", tags: [], verify: [], status: "DOING" },
-      prDir: "/repo/.agentplane/tasks/T-1/pr",
-      metaPath: "/repo/.agentplane/tasks/T-1/pr/meta.json",
-      diffstatPath: "/repo/.agentplane/tasks/T-1/pr/diffstat.txt",
-      verifyLogPath: "/repo/.agentplane/tasks/T-1/pr/verify.log",
-      metaSource: { base: "main", branch: "task/T-1", head_sha: "head-sha" },
-      branch: "task/T-1",
-      base: "main",
-      verifyLogText: "",
-      branchHeadSha: "head-sha",
-      baseHeadSha: "base-sha",
-      changedPaths: [],
-      verifyCommands: [],
-      alreadyVerifiedSha: null,
-      shouldRunVerify: false,
-      protectedBaseRequiresPrMerge: true,
-    });
-    const { cmdIntegrate } = await import("./cmd.js");
-
-    const caught = await cmdIntegrate({
-      cwd: "/repo",
-      taskId: "T-1",
-      mergeStrategy: "squash",
-      runVerify: false,
-      dryRun: false,
-      quiet: false,
-    }).catch((err: unknown) => err);
-
-    expect(caught).toMatchObject({ code: "E_VALIDATION" });
-    expect(String((caught as Error).message)).toContain("GitHub PR base SHA is unavailable");
-    expect(mocks.buildTaskHandoffArtifact).not.toHaveBeenCalled();
-    expect(mocks.writeTaskHandoff).not.toHaveBeenCalled();
-    expect(
-      mocks.execFileAsync.mock.calls.filter((call) => {
-        const args: unknown = (call as unknown[])[1];
-        return Array.isArray(args) && args.includes("PUT");
-      }),
-    ).toEqual([]);
-  });
-
-  it("resolves the protected-base GitHub PR target from live branch state", async () => {
-    mocks.requireOpenGithubPrAtHead.mockResolvedValue({
-      prNumber: 339,
-      prUrl: "https://github.com/example/repo/pull/339",
-      status: "OPEN",
-      mergedAt: null,
-      mergeCommit: null,
-      base: "main",
-      baseSha: "provider-base-sha",
-      headSha: "head-sha",
-    });
-    mocks.prepareIntegrate.mockResolvedValue({
-      ctx: {
-        config: { paths: { workflow_dir: ".agentplane/tasks" } },
-        git: {},
-        taskBackend: {},
-        resolvedProject: { gitRoot: "/repo" },
-      },
-      resolved: { gitRoot: "/repo" },
-      loadedConfig: {
-        workflow_mode: "branch_pr",
-        paths: {
-          worktrees_dir: ".agentplane/worktrees",
-          workflow_dir: ".agentplane/tasks",
-        },
-        commit: { generic_tokens: [] },
-      },
-      task: { id: "T-1", title: "Task", tags: [], verify: [], status: "DOING" },
-      prDir: "/repo/.agentplane/tasks/T-1/pr",
-      metaPath: "/repo/.agentplane/tasks/T-1/pr/meta.json",
-      diffstatPath: "/repo/.agentplane/tasks/T-1/pr/diffstat.txt",
-      verifyLogPath: "/repo/.agentplane/tasks/T-1/pr/verify.log",
-      metaSource: {
-        base: "main",
-        branch: "task/T-1",
-        head_sha: "head-sha",
-      },
-      branch: "task/T-1",
-      base: "main",
-      verifyLogText: "",
-      branchHeadSha: "head-sha",
-      baseHeadSha: "base-sha",
-      changedPaths: [],
-      verifyCommands: [],
-      alreadyVerifiedSha: null,
-      shouldRunVerify: false,
-      protectedBaseRequiresPrMerge: true,
-    });
-    const { cmdIntegrate } = await import("./cmd.js");
-
-    const caught = await cmdIntegrate({
-      cwd: "/repo",
-      taskId: "T-1",
-      mergeStrategy: "squash",
-      runVerify: false,
-      dryRun: false,
-      quiet: false,
-    }).catch((err: unknown) => err);
-
-    expect(caught).toMatchObject({ code: "E_HANDOFF" });
-    expect(mocks.requireOpenGithubPrAtHead).toHaveBeenCalledWith({
-      gitRoot: "/repo",
-      branch: "task/T-1",
-      base: "main",
-      expectedHeadSha: "head-sha",
-      prNumber: null,
-    });
-    expect(mocks.execFileAsync).toHaveBeenCalledWith(
-      "gh",
-      [
-        "api",
-        "-X",
-        "PUT",
-        "repos/example/repo/pulls/339/merge",
-        "-f",
-        "merge_method=rebase",
-        "-f",
-        "sha=head-sha",
-      ],
-      expect.objectContaining({ cwd: "/repo" }),
-    );
-    const handoffCall = mocks.buildTaskHandoffArtifact.mock.calls[0]?.[0] as
-      | { route?: Record<string, unknown> }
-      | undefined;
-    expect(handoffCall?.route).toMatchObject({
-      pr_number: 339,
-      pr_url: "https://github.com/example/repo/pull/339",
-      provider_base_sha: "provider-base-sha",
-    });
-  });
 });
