+++ packages/agentplane/src/commands/shared/route-decision.ts
@@ -41,7 +41,8 @@ import {
   resolveTaskBranchFromContext,
   type CommandContext,
 } from "./task-backend.js";
-import { buildRouteSourceConfidenceBase } from "./source-confidence.js";
+import { buildRouteSourceConfidence } from "./route-decision-source-confidence.js";
+export { buildRouteSourceConfidence } from "./route-decision-source-confidence.js";
 import { hasClosedPreMergeClosureMarker, parsePrMeta } from "./pr-meta.js";
 import { taskCloseAlreadyRecordedOnBase } from "../task/close-tail-state.js";
 import {
@@ -311,33 +312,6 @@ async function resolveLocalTaskBranchFlow(opts: {
     queue: { present: false },
     handoff: { present: false },
     nextAction: `agentplane pr open ${opts.task.id} --author <ROLE>`,
-  };
-}
-
-export function buildRouteSourceConfidence(opts: {
-  remoteEnabled: boolean;
-  remoteResolved: boolean;
-  localDiagnostics: string[];
-}): TaskRouteDecision["sourceConfidence"] {
-  const probeDiagnostics = [...opts.localDiagnostics];
-  if (opts.remoteEnabled && !opts.remoteResolved) {
-    probeDiagnostics.push("remote route probe produced no provider state; local fallback used");
-  }
-  return {
-    ...buildRouteSourceConfidenceBase({
-      remoteEnabled: opts.remoteEnabled,
-      remoteResolved: opts.remoteResolved,
-      batchOwnershipSource: "pr_artifact",
-    }),
-    route_probes: {
-      source: "local_git",
-      freshness: "computed_local",
-      confidence: probeDiagnostics.length > 0 ? "low" : "high",
-      note:
-        probeDiagnostics.length > 0
-          ? probeDiagnostics.join("; ")
-          : "route probes completed without suppressed diagnostics",
-    },
   };
 }
